Update about.html banner, icons and footer
Who Are We: ? This is a question. It's used when seeking information.

Who We Are: This is a statement that identifies or describes a group. ( more accurate to use )

Changed icons, to more suitable options, and improved responsiveness, for the small screens.

Changed font type of footer and minor fixes.


@AkshayTeja